requirements.txt
Django==3.1
djangorestframework==3.11.0
mysqlclient==2.2.4
psycopg2